Added excep to handle offline datalogger
When there is no generation occuring, i.e at night, the datalogging
stick will power down and no longer communicate. To prevent errors, a
try except clause has been added to handle this case. If the stick fails
to respond, it is assumed that no generation is occuring and as such the
current power is set to 0 but the totals are left at their previous
values.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>